Refactor: Move `withOutBodyIface` to new module
We need to give it some additional parameters (the `Context` especially), and
if we don't move it to a separate module first, we'll run into circular module
dependencies (`withOutBodyIface` was defined in `Stream`, which `Context`
imports from).
