+++ Network/HTTP2/H2/Stream.hs
@@ -1,7 +1,4 @@
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE NamedFieldPuns #-}
-{-# LANGUAGE RankNTypes #-}
 
 module Network.HTTP2.H2.Stream where
 
@@ -12,7 +9,6 @@ import Control.Monad
 import Data.IORef
 import Data.Maybe (fromMaybe)
 import Network.Control
-import Network.HTTP.Semantics
 import Network.HTTP.Semantics.IO
 
 import Network.HTTP2.Frame
@@ -99,64 +95,6 @@ closeAllStreams ovar evar mErr = do
     err = Left $ fromMaybe (toException ConnectionIsClosed) mErr
 
 ----------------------------------------------------------------
-
-data StreamTerminated
-    = StreamPushedFinal
-    | StreamCancelled
-    | StreamOutOfScope
-    deriving (Show)
-    deriving anyclass (Exception)
-
-withOutBodyIface
-    :: TBQueue StreamingChunk
-    -> (forall a. IO a -> IO a)
-    -> (OutBodyIface -> IO r)
-    -> IO r
-withOutBodyIface tbq unmask k = do
-    terminated <- newTVarIO Nothing
-    let checkNotTerminated :: STM ()
-        checkNotTerminated = do
-            mTerminated <- readTVar terminated
-            maybe (return ()) throwSTM mTerminated
-
-        iface :: OutBodyIface
-        iface =
-            OutBodyIface
-                { outBodyUnmask = unmask
-                , outBodyPush = \b -> atomically $ do
-                    checkNotTerminated
-                    writeTBQueue tbq $ StreamingBuilder b NotEndOfStream
-                , outBodyPushFinal = \b -> atomically $ do
-                    checkNotTerminated
-                    writeTVar terminated (Just StreamPushedFinal)
-                    writeTBQueue tbq $ StreamingBuilder b (EndOfStream Nothing)
-                    writeTBQueue tbq $ StreamingFinished Nothing
-                , outBodyFlush = atomically $ do
-                    checkNotTerminated
-                    writeTBQueue tbq StreamingFlush
-                , outBodyCancel = \mErr -> atomically $ do
-                    mTerminated <- readTVar terminated
-                    case mTerminated of
-                        Nothing -> do
-                            writeTVar terminated (Just StreamCancelled)
-                            writeTBQueue tbq $ StreamingCancelled mErr
-                        Just _ ->
-                            -- Already terminated
-                            return ()
-                }
-
-        finished :: IO ()
-        finished = atomically $ do
-            mTerminated <- readTVar terminated
-            case mTerminated of
-                Nothing -> do
-                    writeTVar terminated (Just StreamOutOfScope)
-                    writeTBQueue tbq $ StreamingFinished Nothing
-                Just _ ->
-                    -- Already terminated
-                    return ()
-
-    k iface `finally` finished
 
 nextForStreaming
     :: TBQueue StreamingChunk
