+++ http2.cabal
@@ -89,6 +89,7 @@ library
         Network.HTTP2.H2.Context
         Network.HTTP2.H2.EncodeFrame
         Network.HTTP2.H2.HPACK
+        Network.HTTP2.H2.OutBodyIface
         Network.HTTP2.H2.Queue
         Network.HTTP2.H2.Receiver
         Network.HTTP2.H2.Sender
